refactor(runner): split start leaf modules
