+++ crates/runner/src/cmd/start/identity.rs
@@ -1,0 +1,91 @@
+use std::path::Path;
+
+use tracing::info;
+use uuid::Uuid;
+
+use crate::error::{RunnerError, RunnerResult};
+
+/// Load runner ID from `{base_dir}/runner_id`, or generate a new UUID and persist it.
+pub(super) async fn load_or_generate_runner_id(base_dir: &Path) -> RunnerResult<String> {
+    let path = base_dir.join("runner_id");
+    match tokio::fs::read_to_string(&path).await {
+        Ok(contents) => {
+            let id = contents.trim().to_string();
+            Uuid::parse_str(&id).map_err(|e| {
+                RunnerError::Config(format!("invalid runner_id in {}: {e}", path.display()))
+            })?;
+            Ok(id)
+        }
+        Err(e) if e.kind() == std::io::ErrorKind::NotFound => {
+            let id = Uuid::new_v4().to_string();
+            tokio::fs::write(&path, &id).await.map_err(|e| {
+                RunnerError::Config(format!("write runner_id to {}: {e}", path.display()))
+            })?;
+            info!(runner_id = %id, "generated new runner ID");
+            Ok(id)
+        }
+        Err(e) => Err(RunnerError::Config(format!(
+            "read runner_id from {}: {e}",
+            path.display()
+        ))),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[tokio::test]
+    async fn runner_id_generate_and_persist() {
+        let dir = tempfile::tempdir().unwrap();
+        let id1 = load_or_generate_runner_id(dir.path()).await.unwrap();
+        assert!(Uuid::parse_str(&id1).is_ok());
+
+        // Second call reads the same ID
+        let id2 = load_or_generate_runner_id(dir.path()).await.unwrap();
+        assert_eq!(id1, id2);
+    }
+
+    #[tokio::test]
+    async fn runner_id_reads_existing() {
+        let dir = tempfile::tempdir().unwrap();
+        let expected = Uuid::new_v4().to_string();
+        tokio::fs::write(dir.path().join("runner_id"), &expected)
+            .await
+            .unwrap();
+        let id = load_or_generate_runner_id(dir.path()).await.unwrap();
+        assert_eq!(id, expected);
+    }
+
+    #[tokio::test]
+    async fn runner_id_rejects_invalid() {
+        let dir = tempfile::tempdir().unwrap();
+        tokio::fs::write(dir.path().join("runner_id"), "not-a-uuid")
+            .await
+            .unwrap();
+        let result = load_or_generate_runner_id(dir.path()).await;
+        assert!(result.is_err());
+    }
+
+    #[tokio::test]
+    async fn runner_id_trims_whitespace() {
+        let dir = tempfile::tempdir().unwrap();
+        let expected = Uuid::new_v4().to_string();
+        // Write with trailing newline (common with echo/editors)
+        tokio::fs::write(dir.path().join("runner_id"), format!("  {expected}\n"))
+            .await
+            .unwrap();
+        let id = load_or_generate_runner_id(dir.path()).await.unwrap();
+        assert_eq!(id, expected);
+    }
+
+    #[tokio::test]
+    async fn runner_id_rejects_empty_file() {
+        let dir = tempfile::tempdir().unwrap();
+        tokio::fs::write(dir.path().join("runner_id"), "")
+            .await
+            .unwrap();
+        let result = load_or_generate_runner_id(dir.path()).await;
+        assert!(result.is_err());
+    }
+}
