+++ crates/runner/src/cmd/start/mitm_restart.rs
@@ -1,0 +1,190 @@
+use std::time::Duration;
+
+use tracing::{error, info, warn};
+
+use crate::error::RunnerResult;
+use crate::proxy;
+use crate::retry::RetryState;
+
+/// Initial backoff before retrying mitmproxy after a crash.
+pub(super) const MITM_BACKOFF_INITIAL: Duration = Duration::from_secs(1);
+/// Maximum backoff between mitmproxy restart attempts.
+pub(super) const MITM_BACKOFF_MAX: Duration = Duration::from_secs(30);
+/// Stop retrying mitmproxy after this many consecutive failures.
+pub(super) const MITM_MAX_CONSECUTIVE_FAILURES: u32 = 20;
+
+pub(super) type MitmRestartHandle = tokio::task::JoinHandle<RunnerResult<tokio::process::Child>>;
+
+/// Spawn a background mitm restart task when the backoff timer fires
+/// and no restart is already in flight.
+pub(super) async fn maybe_spawn_mitm_restart(
+    mitm: &mut proxy::MitmProxy,
+    crash_rx: &mut tokio::sync::mpsc::Receiver<()>,
+    retry: &mut RetryState<MitmRestartHandle>,
+) {
+    if !retry.timer_ready() {
+        return;
+    }
+    retry.clear_timer();
+    // Defensive drain: clears any crash notification already buffered
+    // from the very crash that armed this restart timer, so it can't
+    // re-trigger the select loop after we're done. `begin_restart` no
+    // longer produces post-call signals — each child owns its own
+    // `stopping` Arc, locked to `true` before kill — so this only
+    // sweeps pre-existing entries.
+    while crash_rx.try_recv().is_ok() {}
+    let params = mitm.begin_restart().await;
+    retry.handle = Some(tokio::spawn(params.spawn()));
+}
+
+/// Handle the result of a background mitm restart task.
+pub(super) fn handle_mitm_restart_result(
+    result: Result<tokio::process::Child, String>,
+    mitm: &mut proxy::MitmProxy,
+    retry: &mut RetryState<MitmRestartHandle>,
+) {
+    match result {
+        Ok(child) => {
+            if retry.consecutive_failures() > 0 {
+                info!(
+                    attempts = retry.consecutive_failures(),
+                    "mitmproxy restarted after failures"
+                );
+            } else {
+                info!("mitmproxy restarted");
+            }
+            mitm.complete_restart(child);
+            retry.on_success();
+        }
+        Err(e) => {
+            // Capture before on_failure() — matches the delay actually scheduled.
+            let next_secs = retry.backoff().as_secs();
+            if !retry.on_failure() {
+                error!(
+                    error = %e,
+                    failures = retry.consecutive_failures(),
+                    "mitmproxy restart abandoned after too many failures"
+                );
+                return;
+            }
+            if retry.consecutive_failures() >= 5 {
+                error!(
+                    error = %e,
+                    failures = retry.consecutive_failures(),
+                    next_attempt_secs = next_secs,
+                    "mitmproxy restart failing persistently"
+                );
+            } else {
+                warn!(
+                    error = %e,
+                    next_attempt_secs = next_secs,
+                    "mitmproxy restart failed"
+                );
+            }
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use std::path::PathBuf;
+
+    use super::*;
+
+    /// Create a MitmProxy for testing (does not start mitmdump).
+    async fn test_mitm() -> (
+        proxy::MitmProxy,
+        tokio::sync::mpsc::Receiver<()>,
+        tempfile::TempDir,
+    ) {
+        let dir = tempfile::tempdir().unwrap();
+        let (mitm, rx) = proxy::MitmProxy::new(proxy::ProxyConfig {
+            mitmdump_bin: PathBuf::from("true"),
+            ca_dir: dir.path().to_path_buf(),
+            addon_dir: dir.path().join("addon"),
+            registry_path: dir.path().join("registry.json"),
+            registry_lock_path: dir.path().join("registry.lock"),
+            api_url: None,
+        })
+        .await
+        .unwrap();
+        (mitm, rx, dir)
+    }
+
+    #[tokio::test]
+    async fn mitm_restart_success_resets_backoff() {
+        let (mut mitm, _rx, _dir) = test_mitm().await;
+        let mut retry: RetryState<MitmRestartHandle> = RetryState::new(
+            MITM_BACKOFF_INITIAL,
+            MITM_BACKOFF_MAX,
+            Some(MITM_MAX_CONSECUTIVE_FAILURES),
+        );
+        retry.backoff = Duration::from_secs(16);
+        retry.consecutive_failures = 5;
+
+        let child = tokio::process::Command::new("true")
+            .stdin(std::process::Stdio::null())
+            .stdout(std::process::Stdio::null())
+            .stderr(std::process::Stdio::null())
+            .spawn()
+            .unwrap();
+
+        handle_mitm_restart_result(Ok(child), &mut mitm, &mut retry);
+
+        assert_eq!(retry.backoff, MITM_BACKOFF_INITIAL);
+        assert_eq!(retry.consecutive_failures, 0);
+    }
+
+    #[tokio::test]
+    async fn mitm_restart_failure_schedules_retry_with_backoff() {
+        let (mut mitm, _rx, _dir) = test_mitm().await;
+        let mut retry: RetryState<MitmRestartHandle> = RetryState::new(
+            MITM_BACKOFF_INITIAL,
+            MITM_BACKOFF_MAX,
+            Some(MITM_MAX_CONSECUTIVE_FAILURES),
+        );
+
+        handle_mitm_restart_result(Err("spawn failed".into()), &mut mitm, &mut retry);
+
+        assert_eq!(retry.consecutive_failures, 1);
+        assert!(retry.restart_at.is_some());
+        assert_eq!(retry.backoff, MITM_BACKOFF_INITIAL * 2);
+    }
+
+    #[tokio::test]
+    async fn mitm_restart_backoff_caps_at_max() {
+        let (mut mitm, _rx, _dir) = test_mitm().await;
+        let mut retry: RetryState<MitmRestartHandle> = RetryState::new(
+            MITM_BACKOFF_INITIAL,
+            MITM_BACKOFF_MAX,
+            Some(MITM_MAX_CONSECUTIVE_FAILURES),
+        );
+        retry.backoff = MITM_BACKOFF_MAX;
+        retry.consecutive_failures = 10;
+
+        handle_mitm_restart_result(Err("spawn failed".into()), &mut mitm, &mut retry);
+
+        assert_eq!(retry.backoff, MITM_BACKOFF_MAX);
+        assert!(retry.restart_at.is_some());
+    }
+
+    #[tokio::test]
+    async fn mitm_restart_circuit_breaker_stops_retrying() {
+        let (mut mitm, _rx, _dir) = test_mitm().await;
+        let mut retry: RetryState<MitmRestartHandle> = RetryState::new(
+            MITM_BACKOFF_INITIAL,
+            MITM_BACKOFF_MAX,
+            Some(MITM_MAX_CONSECUTIVE_FAILURES),
+        );
+        retry.backoff = MITM_BACKOFF_MAX;
+        retry.consecutive_failures = 19;
+
+        handle_mitm_restart_result(Err("binary missing".into()), &mut mitm, &mut retry);
+
+        assert_eq!(retry.consecutive_failures, 20);
+        assert!(
+            retry.restart_at.is_none(),
+            "circuit breaker should prevent further retries"
+        );
+    }
+}
