+++ crates/runner/src/cmd/start/mod.rs
@@ -9,8 +9,6 @@ use tokio::task::JoinSet;
 use tokio_util::sync::CancellationToken;
 use tracing::{debug, error, info, warn};
 
-use uuid::Uuid;
-
 use crate::ids::RunId;
 
 use crate::config::{self, ProfileConfig};
@@ -34,12 +32,20 @@ use crate::status::{RunnerMode, StatusTracker};
 use crate::telemetry::JobTelemetry;
 use crate::types::{ExecutionContext, HeartbeatState, SandboxReuseResult};
 
-/// Initial backoff before retrying mitmproxy after a crash.
-const MITM_BACKOFF_INITIAL: Duration = Duration::from_secs(1);
-/// Maximum backoff between mitmproxy restart attempts.
-const MITM_BACKOFF_MAX: Duration = Duration::from_secs(30);
-/// Stop retrying mitmproxy after this many consecutive failures.
-const MITM_MAX_CONSECUTIVE_FAILURES: u32 = 20;
+mod identity;
+mod mitm_restart;
+mod signals;
+
+use identity::load_or_generate_runner_id;
+use mitm_restart::{
+    MITM_BACKOFF_INITIAL, MITM_BACKOFF_MAX, MITM_MAX_CONSECUTIVE_FAILURES, MitmRestartHandle,
+    handle_mitm_restart_result, maybe_spawn_mitm_restart,
+};
+use signals::{EarlySignals, SignalController};
+
+#[cfg(test)]
+use signals::{handle_drain_signal, handle_resume_signal, handle_stopping_signal};
+
 /// Period between routine heartbeat ticks sent to the server. First
 /// tick is deferred by one period via `interval_at` — see the comment
 /// at the interval construction in `run()`.
@@ -396,8 +402,6 @@ enum SignalSource {
     #[cfg_attr(not(test), allow(dead_code))]
     Override(SignalController),
 }
-
-type MitmRestartHandle = tokio::task::JoinHandle<RunnerResult<tokio::process::Child>>;
 
 async fn run(config: RunConfig) -> RunnerResult<()> {
     let RunConfig {
@@ -1412,253 +1416,6 @@ async fn handle_job_result(
     }
 }
 
-/// Pre-registered signal streams.
-///
-/// Tokio's `signal()` installs the process-wide `sigaction` handler on its
-/// first call per signal kind; before that call, the default disposition
-/// (Term for all four of these) applies. If a drain SIGUSR1 arrives during
-/// startup (e.g. `service install` immediately followed by `service drain`
-/// while the runner is still warming factories), the default action kills
-/// the process and systemd restarts the unit with no one to drain it —
-/// the new PID stays `Running` forever. See issue #10416.
-///
-/// Registering all four streams at the top of [`run_start`] before any
-/// slow work (config load, runtime/factory boot) closes the race: the
-/// sigaction handler is in place and the listener exists, so signals
-/// arriving at any point after registration are queued in the listener's
-/// watch channel and observed by [`SignalController`] as soon as its task
-/// is spawned.
-struct EarlySignals {
-    sigterm: tokio::signal::unix::Signal,
-    sigint: tokio::signal::unix::Signal,
-    sigusr1: tokio::signal::unix::Signal,
-    sigusr2: tokio::signal::unix::Signal,
-}
-
-impl EarlySignals {
-    /// Register the four lifecycle signals (SIGTERM/SIGINT/SIGUSR1/SIGUSR2)
-    /// so they don't fall to their default Term disposition during startup.
-    ///
-    /// Each `signal()` both installs the process-wide `sigaction` handler
-    /// (idempotent via `OnceCell`) and subscribes a fresh watch receiver.
-    /// Bind them via `let` rather than a struct literal so that if a later
-    /// call fails with `?`, the already-subscribed earlier receivers are
-    /// dropped (and unsubscribed) on the error path — obvious at a glance.
-    fn register() -> std::io::Result<Self> {
-        use tokio::signal::unix::{SignalKind, signal};
-        let sigterm = signal(SignalKind::terminate())?;
-        let sigint = signal(SignalKind::interrupt())?;
-        let sigusr1 = signal(SignalKind::user_defined1())?;
-        let sigusr2 = signal(SignalKind::user_defined2())?;
-        Ok(Self {
-            sigterm,
-            sigint,
-            sigusr1,
-            sigusr2,
-        })
-    }
-}
-
-/// Signal-driven mode channel shared between the signal handler task and
-/// the main run loop.
-///
-/// The `RunnerMode` enum is the single source of truth for runner lifecycle
-/// state — `mode_tx` has two writers (the handler for external signals,
-/// and the main loop for the internal Draining → Stopping transition when
-/// `jobs.is_empty()`).
-pub(crate) struct SignalController {
-    pub mode_rx: tokio::sync::watch::Receiver<RunnerMode>,
-    pub mode_tx: tokio::sync::watch::Sender<RunnerMode>,
-    /// Abort handle for the spawned signal-handler task. `None` for test
-    /// overrides where no task was spawned. Teardown calls `.abort()` to
-    /// reap the task symmetrically with `mitm_retry.handle.abort()` — the
-    /// handler otherwise would run until runtime drop, which is safe for
-    /// the runner binary but leaks the task when `run()` is embedded in a
-    /// longer-lived host runtime.
-    pub handler_abort: Option<tokio::task::AbortHandle>,
-}
-
-impl SignalController {
-    /// Spawn the signal-handler task and return a controller handle.
-    ///
-    /// Signal semantics:
-    /// - **SIGUSR1** (drain): from `Running`, send `Draining` (soft,
-    ///   resumable). Ignored from any other state.
-    /// - **SIGUSR2** (resume): from `Draining`, send `Running` (resume normal
-    ///   discovery). Ignored from `Running` / `Stopping` / `Stopped`.
-    /// - **SIGTERM / SIGINT** (hard): send `Stopping`, cancel every in-flight
-    ///   job's token, cancel the discovery token. Bypasses the soft drain
-    ///   so `systemctl stop` exits promptly rather than waiting up to
-    ///   `JOB_TIMEOUT = 2h` for jobs to finish naturally.
-    ///
-    /// ## Race handling
-    ///
-    /// `handle_stopping_signal` sends Stopping **before** locking
-    /// `cancel_tokens`. This ordering lets the main loop close the TOCTOU
-    /// window where a new job is claimed between the handler's iteration
-    /// and its own token insert: the main loop re-reads `mode_rx` after
-    /// insert and sees Stopping via watch's write/read fence.
-    ///
-    /// ## Lifetime
-    ///
-    /// The spawned task loops forever and is implicitly cancelled when the
-    /// tokio runtime is dropped. Its `JoinHandle` is discarded — panics in
-    /// the handler will be logged by tokio but not surfaced.
-    fn spawn(
-        cancel: CancellationToken,
-        cancel_tokens: Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>>,
-        signals: EarlySignals,
-    ) -> Self {
-        let (mode_tx, mode_rx) = tokio::sync::watch::channel(RunnerMode::Running);
-        let tx_for_task = mode_tx.clone();
-        let EarlySignals {
-            mut sigterm,
-            mut sigint,
-            mut sigusr1,
-            mut sigusr2,
-        } = signals;
-        let handle = tokio::spawn(async move {
-            loop {
-                tokio::select! {
-                    _ = sigterm.recv() => {
-                        handle_stopping_signal("SIGTERM", &cancel, &cancel_tokens, &tx_for_task).await;
-                    }
-                    _ = sigint.recv() => {
-                        handle_stopping_signal("SIGINT", &cancel, &cancel_tokens, &tx_for_task).await;
-                    }
-                    _ = sigusr1.recv() => {
-                        handle_drain_signal(&tx_for_task);
-                    }
-                    _ = sigusr2.recv() => {
-                        handle_resume_signal(&tx_for_task);
-                    }
-                }
-            }
-        });
-        Self {
-            mode_rx,
-            mode_tx,
-            handler_abort: Some(handle.abort_handle()),
-        }
-    }
-}
-
-fn handle_drain_signal(mode_tx: &tokio::sync::watch::Sender<RunnerMode>) {
-    let current = *mode_tx.borrow();
-    if current != RunnerMode::Running {
-        warn!(mode = ?current, "SIGUSR1 ignored — only valid from Running");
-        return;
-    }
-    info!("received SIGUSR1, entering Draining (soft drain)");
-    let _ = mode_tx.send(RunnerMode::Draining);
-}
-
-fn handle_resume_signal(mode_tx: &tokio::sync::watch::Sender<RunnerMode>) {
-    let current = *mode_tx.borrow();
-    if current != RunnerMode::Draining {
-        warn!(mode = ?current, "SIGUSR2 ignored — only valid from Draining");
-        return;
-    }
-    info!("received SIGUSR2, resuming to Running");
-    let _ = mode_tx.send(RunnerMode::Running);
-}
-
-async fn handle_stopping_signal(
-    name: &str,
-    cancel: &CancellationToken,
-    cancel_tokens: &Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>>,
-    mode_tx: &tokio::sync::watch::Sender<RunnerMode>,
-) {
-    if *mode_tx.borrow() == RunnerMode::Stopping {
-        warn!(signal = name, "already Stopping, ignoring repeat");
-        return;
-    }
-    info!(signal = name, "initiating hard shutdown");
-    // Send Stopping *before* locking cancel_tokens so that the main loop's
-    // post-insert `mode_rx.borrow()` check catches any job claimed after
-    // our iteration but before send would otherwise publish the state.
-    let _ = mode_tx.send(RunnerMode::Stopping);
-    let tokens = cancel_tokens.lock().await;
-    let count = tokens.len();
-    for (run_id, token) in tokens.iter() {
-        info!(run_id = %run_id, "cancelling active job for hard shutdown");
-        token.cancel();
-    }
-    drop(tokens);
-    info!(active_jobs = count, "dispatched per-job cancellations");
-    cancel.cancel();
-}
-
-/// Spawn a background mitm restart task when the backoff timer fires
-/// and no restart is already in flight.
-async fn maybe_spawn_mitm_restart(
-    mitm: &mut proxy::MitmProxy,
-    crash_rx: &mut tokio::sync::mpsc::Receiver<()>,
-    retry: &mut RetryState<MitmRestartHandle>,
-) {
-    if !retry.timer_ready() {
-        return;
-    }
-    retry.clear_timer();
-    // Defensive drain: clears any crash notification already buffered
-    // from the very crash that armed this restart timer, so it can't
-    // re-trigger the select loop after we're done. `begin_restart` no
-    // longer produces post-call signals — each child owns its own
-    // `stopping` Arc, locked to `true` before kill — so this only
-    // sweeps pre-existing entries.
-    while crash_rx.try_recv().is_ok() {}
-    let params = mitm.begin_restart().await;
-    retry.handle = Some(tokio::spawn(params.spawn()));
-}
-
-/// Handle the result of a background mitm restart task.
-fn handle_mitm_restart_result(
-    result: Result<tokio::process::Child, String>,
-    mitm: &mut proxy::MitmProxy,
-    retry: &mut RetryState<MitmRestartHandle>,
-) {
-    match result {
-        Ok(child) => {
-            if retry.consecutive_failures() > 0 {
-                info!(
-                    attempts = retry.consecutive_failures(),
-                    "mitmproxy restarted after failures"
-                );
-            } else {
-                info!("mitmproxy restarted");
-            }
-            mitm.complete_restart(child);
-            retry.on_success();
-        }
-        Err(e) => {
-            // Capture before on_failure() — matches the delay actually scheduled.
-            let next_secs = retry.backoff().as_secs();
-            if !retry.on_failure() {
-                error!(
-                    error = %e,
-                    failures = retry.consecutive_failures(),
-                    "mitmproxy restart abandoned after too many failures"
-                );
-                return;
-            }
-            if retry.consecutive_failures() >= 5 {
-                error!(
-                    error = %e,
-                    failures = retry.consecutive_failures(),
-                    next_attempt_secs = next_secs,
-                    "mitmproxy restart failing persistently"
-                );
-            } else {
-                warn!(
-                    error = %e,
-                    next_attempt_secs = next_secs,
-                    "mitmproxy restart failed"
-                );
-            }
-        }
-    }
-}
-
 /// References needed to collect and send a heartbeat.
 /// Avoids passing 8+ arguments through `send_heartbeat`.
 struct HeartbeatContext<'a> {
@@ -1745,228 +1502,9 @@ fn collect_heartbeat_state(
     }
 }
 
-/// Load runner ID from `{base_dir}/runner_id`, or generate a new UUID and persist it.
-async fn load_or_generate_runner_id(base_dir: &std::path::Path) -> RunnerResult<String> {
-    let path = base_dir.join("runner_id");
-    match tokio::fs::read_to_string(&path).await {
-        Ok(contents) => {
-            let id = contents.trim().to_string();
-            Uuid::parse_str(&id).map_err(|e| {
-                RunnerError::Config(format!("invalid runner_id in {}: {e}", path.display()))
-            })?;
-            Ok(id)
-        }
-        Err(e) if e.kind() == std::io::ErrorKind::NotFound => {
-            let id = Uuid::new_v4().to_string();
-            tokio::fs::write(&path, &id).await.map_err(|e| {
-                RunnerError::Config(format!("write runner_id to {}: {e}", path.display()))
-            })?;
-            info!(runner_id = %id, "generated new runner ID");
-            Ok(id)
-        }
-        Err(e) => Err(RunnerError::Config(format!(
-            "read runner_id from {}: {e}",
-            path.display()
-        ))),
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    /// Create a MitmProxy for testing (does not start mitmdump).
-    async fn test_mitm() -> (
-        proxy::MitmProxy,
-        tokio::sync::mpsc::Receiver<()>,
-        tempfile::TempDir,
-    ) {
-        let dir = tempfile::tempdir().unwrap();
-        let (mitm, rx) = proxy::MitmProxy::new(proxy::ProxyConfig {
-            mitmdump_bin: PathBuf::from("true"),
-            ca_dir: dir.path().to_path_buf(),
-            addon_dir: dir.path().join("addon"),
-            registry_path: dir.path().join("registry.json"),
-            registry_lock_path: dir.path().join("registry.lock"),
-            api_url: None,
-        })
-        .await
-        .unwrap();
-        (mitm, rx, dir)
-    }
-
-    #[tokio::test]
-    async fn mitm_restart_success_resets_backoff() {
-        let (mut mitm, _rx, _dir) = test_mitm().await;
-        let mut retry: RetryState<MitmRestartHandle> = RetryState::new(
-            MITM_BACKOFF_INITIAL,
-            MITM_BACKOFF_MAX,
-            Some(MITM_MAX_CONSECUTIVE_FAILURES),
-        );
-        retry.backoff = Duration::from_secs(16);
-        retry.consecutive_failures = 5;
-
-        let child = tokio::process::Command::new("true")
-            .stdin(std::process::Stdio::null())
-            .stdout(std::process::Stdio::null())
-            .stderr(std::process::Stdio::null())
-            .spawn()
-            .unwrap();
-
-        handle_mitm_restart_result(Ok(child), &mut mitm, &mut retry);
-
-        assert_eq!(retry.backoff, MITM_BACKOFF_INITIAL);
-        assert_eq!(retry.consecutive_failures, 0);
-    }
-
-    #[tokio::test]
-    async fn mitm_restart_failure_schedules_retry_with_backoff() {
-        let (mut mitm, _rx, _dir) = test_mitm().await;
-        let mut retry: RetryState<MitmRestartHandle> = RetryState::new(
-            MITM_BACKOFF_INITIAL,
-            MITM_BACKOFF_MAX,
-            Some(MITM_MAX_CONSECUTIVE_FAILURES),
-        );
-
-        handle_mitm_restart_result(Err("spawn failed".into()), &mut mitm, &mut retry);
-
-        assert_eq!(retry.consecutive_failures, 1);
-        assert!(retry.restart_at.is_some());
-        assert_eq!(retry.backoff, MITM_BACKOFF_INITIAL * 2);
-    }
-
-    #[tokio::test]
-    async fn mitm_restart_backoff_caps_at_max() {
-        let (mut mitm, _rx, _dir) = test_mitm().await;
-        let mut retry: RetryState<MitmRestartHandle> = RetryState::new(
-            MITM_BACKOFF_INITIAL,
-            MITM_BACKOFF_MAX,
-            Some(MITM_MAX_CONSECUTIVE_FAILURES),
-        );
-        retry.backoff = MITM_BACKOFF_MAX;
-        retry.consecutive_failures = 10;
-
-        handle_mitm_restart_result(Err("spawn failed".into()), &mut mitm, &mut retry);
-
-        assert_eq!(retry.backoff, MITM_BACKOFF_MAX);
-        assert!(retry.restart_at.is_some());
-    }
-
-    #[tokio::test]
-    async fn mitm_restart_circuit_breaker_stops_retrying() {
-        let (mut mitm, _rx, _dir) = test_mitm().await;
-        let mut retry: RetryState<MitmRestartHandle> = RetryState::new(
-            MITM_BACKOFF_INITIAL,
-            MITM_BACKOFF_MAX,
-            Some(MITM_MAX_CONSECUTIVE_FAILURES),
-        );
-        retry.backoff = MITM_BACKOFF_MAX;
-        retry.consecutive_failures = 19;
-
-        handle_mitm_restart_result(Err("binary missing".into()), &mut mitm, &mut retry);
-
-        assert_eq!(retry.consecutive_failures, 20);
-        assert!(
-            retry.restart_at.is_none(),
-            "circuit breaker should prevent further retries"
-        );
-    }
-
-    #[tokio::test]
-    async fn runner_id_generate_and_persist() {
-        let dir = tempfile::tempdir().unwrap();
-        let id1 = load_or_generate_runner_id(dir.path()).await.unwrap();
-        assert!(Uuid::parse_str(&id1).is_ok());
-
-        // Second call reads the same ID
-        let id2 = load_or_generate_runner_id(dir.path()).await.unwrap();
-        assert_eq!(id1, id2);
-    }
-
-    #[tokio::test]
-    async fn runner_id_reads_existing() {
-        let dir = tempfile::tempdir().unwrap();
-        let expected = Uuid::new_v4().to_string();
-        tokio::fs::write(dir.path().join("runner_id"), &expected)
-            .await
-            .unwrap();
-        let id = load_or_generate_runner_id(dir.path()).await.unwrap();
-        assert_eq!(id, expected);
-    }
-
-    #[tokio::test]
-    async fn runner_id_rejects_invalid() {
-        let dir = tempfile::tempdir().unwrap();
-        tokio::fs::write(dir.path().join("runner_id"), "not-a-uuid")
-            .await
-            .unwrap();
-        let result = load_or_generate_runner_id(dir.path()).await;
-        assert!(result.is_err());
-    }
-
-    #[tokio::test]
-    async fn runner_id_trims_whitespace() {
-        let dir = tempfile::tempdir().unwrap();
-        let expected = Uuid::new_v4().to_string();
-        // Write with trailing newline (common with echo/editors)
-        tokio::fs::write(dir.path().join("runner_id"), format!("  {expected}\n"))
-            .await
-            .unwrap();
-        let id = load_or_generate_runner_id(dir.path()).await.unwrap();
-        assert_eq!(id, expected);
-    }
-
-    #[tokio::test]
-    async fn runner_id_rejects_empty_file() {
-        let dir = tempfile::tempdir().unwrap();
-        tokio::fs::write(dir.path().join("runner_id"), "")
-            .await
-            .unwrap();
-        let result = load_or_generate_runner_id(dir.path()).await;
-        assert!(result.is_err());
-    }
-
-    /// Regression test for issue #10416: SIGUSR1 arriving between
-    /// `EarlySignals::register()` and `SignalController::spawn` must still
-    /// drive the mode transition. Previously the handler was registered
-    /// inside the spawned task, so signals delivered during the startup
-    /// window hit the default Term disposition and killed the process.
-    ///
-    /// Timing: raising SIGUSR1 before `spawn()` is deterministic in program
-    /// order, and the outer `timeout(2s)` absorbs the
-    /// `kernel → sigaction → pipe → driver → watch → recv` propagation
-    /// regardless of scheduling — no artificial sleep is needed.
-    ///
-    /// This test raises SIGUSR1 process-wide. It is safe only because no
-    /// other test subscribes to SIGUSR1 — all other tests use
-    /// `SignalSource::Override` and never call `signal()`. If another
-    /// signal-raising test is added, serialize them (e.g. with a
-    /// `Mutex` shared via `OnceLock`) to avoid cross-talk.
-    #[tokio::test]
-    async fn signal_buffered_before_spawn_is_delivered() {
-        let signals = EarlySignals::register().expect("register");
-
-        nix::sys::signal::raise(nix::sys::signal::Signal::SIGUSR1).expect("raise");
-
-        let controller = SignalController::spawn(
-            CancellationToken::new(),
-            Arc::new(tokio::sync::Mutex::new(HashMap::new())),
-            signals,
-        );
-        let mut mode_rx = controller.mode_rx;
-
-        tokio::time::timeout(Duration::from_secs(2), mode_rx.changed())
-            .await
-            .expect("buffered SIGUSR1 should drive mode within 2s")
-            .expect("mode channel closed");
-        assert_eq!(*mode_rx.borrow(), RunnerMode::Draining);
-
-        // Abort so the task releases its Signal stream subscriptions
-        // and does not linger to consume signals raised by later tests.
-        if let Some(abort) = controller.handler_abort {
-            abort.abort();
-        }
-    }
 
     // -----------------------------------------------------------------------
     // collect_heartbeat_state: running_count excludes idle VMs
@@ -3007,89 +2545,6 @@ mod tests {
         );
 
         shutdown(&env, run_handle).await;
-    }
-
-    /// `handle_drain_signal` state guard: SIGUSR1 is honored only from
-    /// Running. Mirrors `handle_resume_signal`'s Draining-only guard.
-    #[test]
-    fn drain_signal_state_guards() {
-        // Running → Draining (sanity: the one legal transition).
-        let (tx, _rx) = tokio::sync::watch::channel(RunnerMode::Running);
-        handle_drain_signal(&tx);
-        assert_eq!(*tx.borrow(), RunnerMode::Draining);
-
-        // Draining → ignored (no self-transition).
-        let (tx, _rx) = tokio::sync::watch::channel(RunnerMode::Draining);
-        handle_drain_signal(&tx);
-        assert_eq!(*tx.borrow(), RunnerMode::Draining);
-
-        // Stopping → ignored (cannot reverse teardown).
-        let (tx, _rx) = tokio::sync::watch::channel(RunnerMode::Stopping);
-        handle_drain_signal(&tx);
-        assert_eq!(*tx.borrow(), RunnerMode::Stopping);
-
-        // Stopped → ignored (runner has exited its loop).
-        let (tx, _rx) = tokio::sync::watch::channel(RunnerMode::Stopped);
-        handle_drain_signal(&tx);
-        assert_eq!(*tx.borrow(), RunnerMode::Stopped);
-    }
-
-    /// `handle_resume_signal` state guard: SIGUSR2 is honored only from
-    /// Draining. The integration test `resume_after_stopping_is_ignored`
-    /// covers the Stopping case; this pins the full matrix as a unit test.
-    #[test]
-    fn resume_signal_state_guards() {
-        // Draining → Running (sanity: the one legal transition).
-        let (tx, _rx) = tokio::sync::watch::channel(RunnerMode::Draining);
-        handle_resume_signal(&tx);
-        assert_eq!(*tx.borrow(), RunnerMode::Running);
-
-        // Running → ignored (nothing to resume from).
-        let (tx, _rx) = tokio::sync::watch::channel(RunnerMode::Running);
-        handle_resume_signal(&tx);
-        assert_eq!(*tx.borrow(), RunnerMode::Running);
-
-        // Stopping → ignored (too late).
-        let (tx, _rx) = tokio::sync::watch::channel(RunnerMode::Stopping);
-        handle_resume_signal(&tx);
-        assert_eq!(*tx.borrow(), RunnerMode::Stopping);
-
-        // Stopped → ignored.
-        let (tx, _rx) = tokio::sync::watch::channel(RunnerMode::Stopped);
-        handle_resume_signal(&tx);
-        assert_eq!(*tx.borrow(), RunnerMode::Stopped);
-    }
-
-    /// `handle_stopping_signal` idempotency: a repeat invocation takes the
-    /// "already Stopping" guard and returns without re-iterating
-    /// `cancel_tokens` or re-cancelling `cancel`.
-    #[tokio::test]
-    async fn stopping_signal_repeat_is_idempotent() {
-        let (tx, _rx) = tokio::sync::watch::channel(RunnerMode::Running);
-        let cancel = CancellationToken::new();
-        let tokens: Arc<tokio::sync::Mutex<HashMap<RunId, CancellationToken>>> =
-            Arc::new(tokio::sync::Mutex::new(HashMap::new()));
-
-        // First call: transitions, cancels main cancel.
-        handle_stopping_signal("SIGTERM", &cancel, &tokens, &tx).await;
-        assert_eq!(*tx.borrow(), RunnerMode::Stopping);
-        assert!(cancel.is_cancelled());
-
-        // Insert a sentinel token *after* the first call so we can prove
-        // the repeat did not re-iterate the map.
-        let sentinel = CancellationToken::new();
-        tokens
-            .lock()
-            .await
-            .insert(RunId::new_v4(), sentinel.clone());
-
-        // Repeat call: must early-return on the already-Stopping guard.
-        handle_stopping_signal("SIGTERM", &cancel, &tokens, &tx).await;
-        assert_eq!(*tx.borrow(), RunnerMode::Stopping);
-        assert!(
-            !sentinel.is_cancelled(),
-            "repeat must not re-iterate cancel_tokens and cancel late-inserted sentinel",
-        );
     }
 
     /// Draining auto-transitions to Stopping when jobs drain naturally.
